API

Function Documentation

Modelling

Turing.@model — Macro.

@model(name, fbody)

Wrapper for models.

Usage:

@model model() = begin
  # body
end

Example:

@model gauss() = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0,sqrt.(s))
  1.5 ~ Normal(m, sqrt.(s))
  2.0 ~ Normal(m, sqrt.(s))
  return(s, m)
end

source

Turing.@~ — Macro.

var_name ~ Distribution()

Tilde notation ~ can be used to specifiy a variable follows a distributions.

If var_name is an un-defined variable or a container (e.g. Vector or Matrix), this variable will be treated as model parameter; otherwise if var_name is defined, this variable will be treated as data.

source

Samplers

Turing.Sampler — Type.

Sampler{T}

Generic interface for implementing inference algorithms. An implementation of an algorithm should include the following:

A type specifying the algorithm and its parameters, derived from InferenceAlgorithm

A method of sample function that produces results of inference, which is where actual inference happens.

Turing translates models to chunks that call the modelling functions at specified points. The dispatch is based on the value of a sampler variable. To include a new inference algorithm implements the requirements mentioned above in a separate file, then include that file at the end of this one.

source

Turing.Gibbs — Type.

Gibbs(n_iters, alg_1, alg_2)

Compositional MCMC interface.

Usage:

alg = Gibbs(1000, HMC(1, 0.2, 3, :v1), PG(20, 1, :v2))

source

Turing.HMC — Type.

HMC(n_iters::Int, epsilon::Float64, tau::Int)

Hamiltonian Monte Carlo sampler.

Usage:

HMC(1000, 0.05, 10)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0,sqrt.(s))
  x[1] ~ Normal(m, sqrt.(s))
  x[2] ~ Normal(m, sqrt.(s))
  return s, m
end

sample(gdemo([1.5, 2]), HMC(1000, 0.05, 10))

source

Turing.HMCDA — Type.

HMCDA(n_iters::Int, n_adapt::Int, delta::Float64, lambda::Float64)

Hamiltonian Monte Carlo sampler wiht Dual Averaging algorithm.

Usage:

HMCDA(1000, 200, 0.65, 0.3)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0,sqrt.(s))
  x[1] ~ Normal(m, sqrt.(s))
  x[2] ~ Normal(m, sqrt.(s))
  return s, m
end

sample(gdemo([1.5, 2]), HMCDA(1000, 200, 0.65, 0.3))

source

Turing.IPMCMC — Type.

IPMCMC(n_particles::Int, n_iters::Int, n_nodes::Int, n_csmc_nodes::Int)

Particle Gibbs sampler.

Usage:

IPMCMC(100, 100, 4, 2)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0,sqrt(s))
  x[1] ~ Normal(m, sqrt(s))
  x[2] ~ Normal(m, sqrt(s))
  return s, m
end

sample(gdemo([1.5, 2]), IPMCMC(100, 100, 4, 2))

source

Turing.IS — Type.

IS(n_particles::Int)

Importance sampling algorithm object.

n_particles is the number of particles to use

Usage:

IS(1000)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0,sqrt.(s))
  x[1] ~ Normal(m, sqrt.(s))
  x[2] ~ Normal(m, sqrt.(s))
  return s, m
end

sample(gdemo([1.5, 2]), IS(1000))

source

Turing.MH — Type.

MH(n_iters::Int)

Metropolis-Hasting sampler.

Usage:

MH(100, (:m, (x) -> Normal(x, 0.1)))

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0,sqrt(s))
  x[1] ~ Normal(m, sqrt(s))
  x[2] ~ Normal(m, sqrt(s))
  return s, m
end

sample(gdemo([1.5, 2]), MH(1000, (:m, (x) -> Normal(x, 0.1)), :s)))

source

Turing.NUTS — Type.

NUTS(n_iters::Int, n_adapt::Int, delta::Float64)

No-U-Turn Sampler (NUTS) sampler.

Usage:

NUTS(1000, 200, 0.6j_max)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0,sqrt.(s))
  x[1] ~ Normal(m, sqrt.(s))
  x[2] ~ Normal(m, sqrt.(s))
  return s, m
end

sample(gdemo([1.j_max, 2]), NUTS(1000, 200, 0.6j_max))

source

Turing.PG — Type.

PG(n_particles::Int, n_iters::Int)

Particle Gibbs sampler.

Usage:

PG(100, 100)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0,sqrt.(s))
  x[1] ~ Normal(m, sqrt.(s))
  x[2] ~ Normal(m, sqrt.(s))
  return s, m
end

sample(gdemo([1.5, 2]), PG(100, 100))

source

Turing.PMMH — Type.

PMMH(n_iters::Int, smc_alg:::SMC, parameters_algs::Tuple{MH})

Particle independant Metropolis–Hastings and Particle marginal Metropolis–Hastings samplers.

Usage:

alg = PMMH(100, SMC(20, :v1), MH(1,:v2))
alg = PMMH(100, SMC(20, :v1), MH(1,(:v2, (x) -> Normal(x, 1))))

source

Turing.SGHMC — Type.

SGHMC(n_iters::Int, learning_rate::Float64, momentum_decay::Float64)

Stochastic Gradient Hamiltonian Monte Carlo sampler.

Usage:

SGHMC(1000, 0.01, 0.1)

Example:

@model example begin
  ...
end

sample(example, SGHMC(1000, 0.01, 0.1))

source

Turing.SGLD — Type.

SGLD(n_iters::Int, step_size::Float64)

Stochastic Gradient Langevin Dynamics sampler.

Usage:

SGLD(1000, 0.5)

Example:

@model example begin
  ...
end

sample(example, SGLD(1000, 0.5))

source

Turing.SMC — Type.

SMC(n_particles::Int)

Sequential Monte Carlo sampler.

Usage:

SMC(1000)

Example:

# Define a simple Normal model with unknown mean and variance.
@model gdemo(x) = begin
  s ~ InverseGamma(2,3)
  m ~ Normal(0,sqrt.(s))
  x[1] ~ Normal(m, sqrt.(s))
  x[2] ~ Normal(m, sqrt.(s))
  return s, m
end

sample(gdemo([1.5, 2]), SMC(1000))

source

Index

Turing.Gibbs

Turing.HMC

Turing.HMCDA

Turing.IPMCMC

Turing.IS

Turing.MH

Turing.NUTS

Turing.PG

Turing.PMMH

Turing.SGHMC

Turing.SGLD

Turing.SMC

Turing.Sampler

Turing.@model

Turing.@~